+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.4.0 - 2017-09-01
 - Project - external_organisations, funded.
-- Publication - associated.
+- Publication - associated, external_organisations.
 
 ## 1.3.0 - 2017-04-03
 ### Added